Remove median as backup in Wt_Sampled_3
sum of weight from LW relationship will not be NA if there are fish without
weights b/c na.rm is in the sum function. therefore, when you won't get a
median for any sample_no that doesn't also give you a sum.

code was also simplified using split and unsplit to get the sum by sample_no
rather than finding it for the tow data frame and then searching for
matching rows in the Pdata data frame.