Fix bug in MALES_WGT + FEMALES_WGT when one or other is NA b/c no samples
I assume that if MALES_NUM is not NA then there should be a weight in MALES_WGT
and vice versa for females. When there is no species-specific sample size,
the code assumes the weight is zero instead of NA. This allows a sum
to be calculated of 0 + the sex with a weight rather than NA.
NA is still generated when there is a non-NA sex-specific sample size.
Then the weight-length relationship will need to be used.