Stop relying on tows and sample_no:cluster for Sampled_2, work with Pdata instead.
Get rid of code that was checking what Andy did but was commented out.
Use unlist and split instead of creating new dataframes and then
indexing the appropriate columns to expand the information back up.
I tested this to make sure the answers were the same for sablefish. Less
chance for errors when working with a single object than back indexing.