Wt_Sampled_1 should include fish with unkown sex.
Had to create a workaround for dmp files that don't yet have the unk_wgt
and unk_num columns, which are available in PacFIN but only from a special
table.
Got rid of use of tows table and working straight with Pdata.
Will need to think about the times when you don't want to include
unsexed fish in your expansion, or expand them separately. Which could be
done by creating a unique sample number.