fix: variable value.names now explicitly specified in rescale and round
value.names was specified earlier within the function but not
explicitly passed to rescale.comps or round.comps.
Also, changed to using grep and value = TRUE rather than an artificial
L and A vector. This allows for comps to be labeled with F, L, M, or A
of upper or lowercase.