docs: begin readme through cleanPacFIN
Documentation is needed for reviewers of U.S. West Coast stock assessments.
This readme will serve as the current source of information in how the code
works and how expansions are done in general for PacFIN data.
The current commit documents the cleaning stages of the code.
Next will be the actual expansions. Comments are welcome and suggested ways
to display house this information rather than the readme are welcome as well.