refactor(createSampleTable): change function name createSampleTable tableSample
Change the name of createSampleTable to tableSample to align with the package
naming convention of plot... and now table... Users will know which functions
generate plots and which generate tables using the prefixes and having plot
and table first allows alphabetized listings that group like functions together.

Export the function adding it to the namespace.