refactor(get_Expansion_2 Units): use match.arg with defaults in function call
By default match.arg will take the vector of values from the call in function()
so you don't need to specify them again in the call to match.arg. Also by using
several.ok = FALSE only one value is allowed and the first value in the default
vector will be used. So, for this function that defaults to "MT". I also increased
the amount of documentation for the old deprecated argument to make it clear how
TRUE leads to MT.