feat: start images directory with marine_areas_all from WDFW
Add directory for images, i.e., figures that were not created by our team
and without associated code, that could be useful in the future.
Captions are currently included in the readme though we could change
this information to be in files with the same name as the image in the future.