fix: calculate n tows based on both M and F combined
The old approach calculated the number of unique tows for each sex. However, if there are sexed fish, the most likely way these data are used in SS3 is within a sexed combined (sex = 3) data frame which should have the sample size calculated based on males and females combined rather than separate. The all tows column which has the number of tows regardless of sex could be an overestimate of the unique number of tows if there is a good mixture of unsexed and sexed fish in the composition data.