fix: add calculations for samps and tows where fish was sexed
This commit add the calculation of samps and tows when fished were sexed (e.g., female or male). The columns associated with these values are called "both" or "btows". This revision was needed particularly for length and marginal age compositions since sexed fish are nearly always input to SS3 in one matrix with sex = 3 so the effective sample size should be calculated for tows with sexed fish. The old approach either had sexed and unsexed calculated separately with a alltows column that was all unique tows. This revision will be more important for species with a large number of unsexed fish compared to sexed.