fix(PullBDS.PacFIN) pivot_wider(names_sort = TRUE)
Uses names_sort = TRUE, where the default is FALSE, to ensure that
column names generated by pivot_wider with respect to multiple age
reads are in the same order every time. Without sort = TRUE, then
the order of the column names is dependent on the order of the
entries in names_from, that is if AGE_SEQUENCE_NUMBER 3 is a row
ahead of 2 then the resulting columns in the returned data frame
will be age3 then age2. Found this will working with petrale sole.