feat(PullBDS.PacFIN): New column FISH_WEIGHT_GUTTED
Thanks to Kristen Hinton for letting me know that sometimes in WA
purposive samples can be measured twice for weight. Instead of leaving
these samples as two rows, I made a new column and used pivot_wider()
to place the smaller weighted in FISH_WEIGHT_GUTTED for when they
are experimenting with weights of headed and gutted fish versus whole
fish.

All duplicate records are removed using the same functionality
implemented above as a byproduct.

Close #105