doc(Pull*): use inheritParams
Stop using template for password and username because the best practices
are to use @inheritParams. This allows users to see the default text in
a typical R file rather than having to navigate to man/roxygen/templates
or wherever the templates are stored. This will be the first commit in a
series of commits to change from old practices to new practices.