fix(getDB): bug in default argument username
The default for `username = getUserName(database)`, which would have led
to an error because R would not know what the `database` object is
because it is another argument. You cannot call an argument in the
setting of another argument when you are evaluating a function. This is
now moved to only if the argument `username` is missing. Documentation
for the input argument has also been fixed.

Additional updates were made to the to documentation and a message that
was hard wired to PacFIN and now it uses a call to paste in combination
with the user supplied database argument.