fix(writeComps): Creates an SS3 file
Removed the loops in writeComps in favor of using dplyr and tried
to align the column names to match what nwfscSurvey is outputting to
make accounting easier.

Added an example worflow for sablefish.

Still need to work on sample size and ensure that the correct column
is being used for the composition data. Thank you everyone for your
patience.