fix: replace Inf records with 1
There can be records that cannot be expanded due to missing sample weights.  This sets the expansions for those records to 1 similar to records with NAs (missing trip weights). This primarily impacts WA records which often do not have EXP_WT or RWT_LBS.  Since this has a large impact on the expanded composition data, this change should be reviewed carefully.