refactor: add changes from other branch and update messages
Include the minor changes made in the expansion2_error_fix branch here which makes the code robust to whether the Catch file is a data.frame or a tibble.  Additionally, the messages have been changed to cli_alert_danger for years that are not included in the Catch file or the Catch file has catch of 0 when there are bds data.